+++ src/Standalone/sphere.h
@@ -6,13 +6,13 @@
 class FSphere : public FHittable
 {
 public:
-    FSphere(const FPoint3& CenterIn, float RadiusIn, std::shared_ptr<FMaterial> MaterialIn);
+    FSphere(const FPoint3& CenterIn, double RadiusIn, std::shared_ptr<FMaterial> MaterialIn);
 
     bool Hit(const FRay& Ray, FInterval Interval, FHitRecord& HitRecordOut) const override;
 
 private:
     FPoint3 Center;
-    float Radius;
+    double Radius;
     std::shared_ptr<FMaterial> Material;
 };
